requirements.txt
cachetools
certifi
charset-normalizer
face_detection
filelock
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
Jinja2
joblib
lz4
mpmath
mtcnn
networkx
numpy
oauthlib
opencv-contrib-python
opencv-stubs
pandas
pillow
proto-plus
protobuf
pyasn1
pyasn1_modules
pyparsing
python-dateutil
pytz
requests
requests-oauthlib
rsa
six
sympy
torchvision
torchvision
#triton
typing_extensions
tzdata
uritemplate
urllib3
